rename model => interface
